refactor(pa): Share helpers via utils.py and trim dead code
- Move shared int-arith (cdiv, pow2 div/rem), device-math (rcp,
  exp2, maxnumf) and state-reshape (unflatten_k) helpers into
  kernels/utils.py; dedup across pa_decode_fp8 / pa_decode_swa / pa_metadata
  (pa_decode_swa keeps its own fly_math-based exp2_f32_fast)
- Drop dead code in pa_decode_fp8 left by the metadata move (unused
  constants/helpers, the _load_*_scales helpers, orphaned _make_pa_phase_helpers
  and _build_pa_thread_invariants params)
- Collapse redundant cache_scale_vecs flag into per_token_kv
- Trim verbose/redundant comments
